moved spotify to alerts/spotify
