+++ src/main/java/de/slimecloud/slimeball/features/alerts/spotify/Spotify.java
@@ -1,4 +1,4 @@
-package de.slimecloud.slimeball.features.alerts;
+package de.slimecloud.slimeball.features.alerts.spotify;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.hc.core5.http.ParseException;
