+++ src/main/java/de/slimecloud/slimeball/main/SlimeBot.java
@@ -13,8 +13,8 @@ import de.slimecloud.slimeball.config.GuildConfig;
 import de.slimecloud.slimeball.config.LogForwarding;
 import de.slimecloud.slimeball.config.commands.ConfigCommand;
 import de.slimecloud.slimeball.features.alerts.HolidayAlert;
-import de.slimecloud.slimeball.features.alerts.Spotify;
-import de.slimecloud.slimeball.features.alerts.SpotifyAlert;
+import de.slimecloud.slimeball.features.alerts.spotify.Spotify;
+import de.slimecloud.slimeball.features.alerts.spotify.SpotifyAlert;
 import de.slimecloud.slimeball.features.birthday.Birthday;
 import de.slimecloud.slimeball.features.birthday.BirthdayAlert;
 import de.slimecloud.slimeball.features.birthday.BirthdayListener;
